Clean script and Visual Feedback on Error
Clean script includes h5 as well as img.
Though images don't have any text for aeneas to work with there may be a way to assign some portion of audio highlighting to the images.

Errors are now much more visually noticable. Turning the feed box as well as the upload box red. Stating that its a error in aeneas. Most likely its a mismatch between the number of mp3 files and segmented files. (Headings or mp3 need manual fixing)