Daisy to Epub3
Using the Daisy Pipeline we can now convert whole books into epub3 format.

This should speedup the process of converting a large collection of daisy books into epub3 books when needed in a few years.

dp2.exe and config.yml are located in the root folder for easy access to the pipeline commandline

rimraf dependency added to remove recursively folders and files uploaded to the server.

Most of the code used is simply a copy and paste of the smil creation, some refactoring may be allotted in the future.